+++ extension/src/Settings/LanguageFunctionsSettings.ts
@@ -8,7 +8,7 @@ import { Settings } from "./Settings";
 export class LanguageFunctionsSettings {
   translationMode: TranslationMode;
   useExternalTranslationTool: boolean;
-  setExactMatchToState: TargetState;
+  setExactMatchToState?: TargetState;
   clearTargetWhenSourceHasChanged: boolean;
   searchOnlyXlfFiles: boolean;
   detectInvalidValuesEnabled: boolean;
@@ -18,10 +18,8 @@ export class LanguageFunctionsSettings {
   useMatchingSetting: boolean;
   replaceSelfClosingXlfTags: boolean;
   searchReplaceBeforeSaveXliff: ISearchReplaceBeforeSaveXliff[] = [];
-  exactMatchState?: TargetState;
   formatXml = true;
   refreshXlfAfterFindNextUntranslated: boolean;
-  useDictionaryInDTSImport: boolean;
   preferLockedTranslations: boolean;
   ignoreMissingTransUnitsOnImport: boolean;
   importTranslationWithDifferentSource: boolean;
@@ -42,10 +40,8 @@ export class LanguageFunctionsSettings {
     this.useMatchingSetting = settings.matchTranslation;
     this.replaceSelfClosingXlfTags = settings.replaceSelfClosingXlfTags;
     this.searchReplaceBeforeSaveXliff = settings.searchReplaceBeforeSaveXliff;
-    this.exactMatchState = this.getDtsExactMatchToState(settings);
     this.refreshXlfAfterFindNextUntranslated =
       settings.refreshXlfAfterFindNextUntranslated;
-    this.useDictionaryInDTSImport = settings.useDictionaryInDTSImport;
     this.preferLockedTranslations = settings.preferLockedTranslations;
     this.ignoreMissingTransUnitsOnImport =
       settings.ignoreMissingTransUnitsOnImport;
@@ -53,32 +49,12 @@ export class LanguageFunctionsSettings {
       settings.importTranslationWithDifferentSource;
   }
 
-  private getDtsExactMatchToState(settings: Settings): TargetState | undefined {
-    const setDtsExactMatchToState: string = settings.setDtsExactMatchToState;
-    let exactMatchState: TargetState | undefined;
-    if (setDtsExactMatchToState.toLowerCase() !== "(keep)") {
-      exactMatchState = setDtsExactMatchToState as TargetState;
-    }
-    return exactMatchState;
-  }
-
   private getTranslationMode(settings: Settings): TranslationMode {
-    const useDTS: boolean = settings.useDTS;
-    if (useDTS) {
-      return TranslationMode.dts;
-    }
     const useExternalTranslationTool: boolean =
       settings.useExternalTranslationTool;
     if (useExternalTranslationTool) {
       return TranslationMode.external;
     }
     return TranslationMode.nabTags;
-  }
-
-  public get useDTSDictionary(): boolean {
-    return (
-      this.translationMode === TranslationMode.dts &&
-      this.useDictionaryInDTSImport
-    );
   }
 }